requirements.txt
streamlit
deepgram-sdk
groq
langchain
langchain-groq
langchain-openai
python-dotenv
faiss-cpu
asyncio
langchain-community
pypdf
nest_asyncio
requests
certifi
pyaudio
wave
pydub
simpleaudio
sounddevice